Began moving QueryType, QueryResult, QueryResultBuilder, Status
