+++ surrealdb/Cargo.toml
@@ -16,16 +16,16 @@ license-file.workspace = true
 [features]
 # Public features
 default = ["protocol-ws", "rustls"]
-allocator = ["surrealdb-core/allocator"]
+allocator = ["dep:surrealdb-core", "surrealdb-core/allocator"]
 protocol-http = ["dep:reqwest"]
 protocol-ws = ["dep:tokio-tungstenite", "tokio/time"]
-kv-mem = ["surrealdb-core/kv-mem", "tokio/time"]
-kv-indxdb = ["surrealdb-core/kv-indxdb"]
-kv-rocksdb = ["surrealdb-core/kv-rocksdb", "tokio/time"]
-kv-tikv = ["surrealdb-core/kv-tikv", "tokio/time"]
-kv-surrealkv = ["surrealdb-core/kv-surrealkv", "tokio/time"]
+kv-mem = ["dep:surrealdb-core", "surrealdb-core/kv-mem", "tokio/time"]
+kv-indxdb = ["dep:surrealdb-core", "surrealdb-core/kv-indxdb"]
+kv-rocksdb = ["dep:surrealdb-core", "surrealdb-core/kv-rocksdb", "tokio/time"]
+kv-tikv = ["dep:surrealdb-core", "surrealdb-core/kv-tikv", "tokio/time"]
+kv-surrealkv = ["dep:surrealdb-core", "surrealdb-core/kv-surrealkv", "tokio/time"]
 scripting = ["surrealdb-core/scripting"]
-http = ["surrealdb-core/http"]
+http = ["dep:surrealdb-core", "surrealdb-core/http"]
 native-tls = [
     "dep:native-tls",
     "reqwest?/native-tls",
@@ -37,9 +37,9 @@ rustls = [
     "tokio-tungstenite?/rustls-tls-webpki-roots",
 ]
 ml = ["surrealdb-core/ml"]
-jwks = ["surrealdb-core/jwks"]
-arbitrary = ["surrealdb-core/arbitrary"]
-allocation-tracking = ["surrealdb-core/allocation-tracking"]
+jwks = ["dep:surrealdb-core", "surrealdb-core/jwks"]
+arbitrary = ["dep:surrealdb-core", "surrealdb-core/arbitrary"]
+allocation-tracking = ["dep:surrealdb-core", "surrealdb-core/allocation-tracking"]
 
 [package.metadata.docs.rs]
 rustdoc-args = ["--cfg", "docsrs"]
@@ -57,7 +57,8 @@ targets = []
 
 [dependencies]
 # workspace internal dependencies
-surrealdb-core = { workspace = true, default-features = false }
+surrealdb-core = { workspace = true, default-features = false, optional = true }
+surrealdb-client-core = { workspace = true }
 surrealdb-types.workspace = true
 surrealdb-types-derive = { workspace = true, features = ["sdk-path"] }
