+++ surrealdb/src/method/query.rs
@@ -8,7 +8,7 @@ use futures::StreamExt;
 use futures::future::Either;
 use futures::stream::SelectAll;
 use indexmap::IndexMap;
-use surrealdb_core::dbs::QueryType;
+use surrealdb_client_core::response::QueryType;
 use surrealdb_core::rpc::DbResultStats;
 use surrealdb_types::Error as TypesError;
 use uuid::Uuid;
